Requirements 4 - 6
Users can specify their next move, which also publicly displays the
board in the channel after the move with a reminder of whose turn it is.
Only the user whose turn it is can make the next move.
When a turn is taken that ends the game, the response indicates this
along with who won.